ArimaKochi.glyphs Fix production names
“there is an error in the GlyphData. The production name of the
aulengthmark-malayalam and rakar-malayalam were switched. You need to
fix this either in a custom glyphdata.xml or directly in the file.
Select the aulengthmark-malayalam, hit cmd+opt+I, check the button at
the production name and assign uni0D57 as value. The same for the
rakar-malayalam, assign uni0D4D0D30.”